% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLS-package.R, R/ConsensusOPLS.R
\docType{package}
\name{ConsensusOPLS}
\alias{ConsensusOPLS}
\alias{_PACKAGE}
\alias{ConsensusOPLS-package}
\title{ConsensusOPLS}
\usage{
ConsensusOPLS(
  data,
  Y,
  maxPcomp = 1,
  maxOcomp = 5,
  modelType = "da",
  nperm = 100,
  cvType = "nfold",
  nfold = 5,
  nMC = 100,
  cvFrac = 4/5,
  mc.cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{A list of data blocks.}

\item{Y}{A vector, factor, dummy matrix or numerical matrix for the response.}

\item{maxPcomp}{Maximum number of Y-predictive components. Default, 1.}

\item{maxOcomp}{Maximum number of Y-orthogonal components. Default, 5.}

\item{modelType}{Type of OPLS regression model, either `reg` for regression or 
`da` for discriminant analysis. Default, `da`.}

\item{nperm}{Number of random permutations. Default, 100.}

\item{cvType}{Type of cross-validation used. Either `nfold` for n-fold cross-validation, 
where \code{nfold} is look up, or `mccv` for Monte Carlo CV, or `mccvb` 
for Monte Carlo class-balanced cross-validation, where \code{nMC} and \code{cvFrac} are used.
Default, `nfold`.}

\item{nfold}{Number of folds performed in n-fold cross-validation. Default, 5.}

\item{nMC}{An integer indicating the number of rounds performed when cvType is `mccv` or `mccvb`.
Default, 100.}

\item{cvFrac}{A numeric value indicating the fraction of observations from \code{data} 
used in the training set for `mccv` or `mccvb` cross-validation. Default, 4/5.}

\item{mc.cores}{Number of cores for parallel computing. Default: 1.}

\item{verbose}{A logical indicating if the computation progress will be shown. Default, FALSE.}
}
\value{
\code{ConsensusOPLS} returns a list of 
\item{\code{optModel}}{ optimal consensus OPLS model}
\item{\code{permModels}}{ models with permuted responses}
\item{\code{permStats}}{ permutation statistics}
\item{\code{plots}}{ plots}
}
\description{
A short description...

This package uses functions from the K-OPLS package, developed by Max Bylesjo, 
University of Umea, Judy Fonville and Mattias Rantalainen, Imperial College.

Copyright (c) 2007-2010 Max Bylesjo, Judy Fonville and Mattias Rantalainen 

This code has been extended and adapted under the terms of the GNU General 
Public License version 2 as published by the Free Software Foundation.

Constructs the consensus OPLS model with an optimal number of orthogonal components
for given data blocks and response, and evaluate the model quality w.r.t other models 
built with randomly permuted responses.
}
\examples{
data(demo_3_Omics)
res <- ConsensusOPLS(data=demo_3_Omics[c("MetaboData", "MicroData", "ProteoData")], 
                     Y=demo_3_Omics$Y,
                     maxPcomp=1, maxOcomp=2, 
                     modelType='da',
                     nperm=5)
str(res)
}
\author{
\strong{Maintainer}: Van Du T. Tran \email{thuong.tran@sib.swiss} (\href{https://orcid.org/0000-0003-2074-5029}{ORCID})

Authors:
\itemize{
  \item Celine Bougel \email{celine.bougel@laposte.net} (\href{https://orcid.org/0000-0002-1598-7827}{ORCID})
  \item Julien Boccard \email{julien.boccard@unige.ch} (\href{https://orcid.org/0000-0001-5913-9566}{ORCID})
  \item Florence Mehl \email{florence.mehl@sib.swiss} (\href{https://orcid.org/0000-0002-9619-1707}{ORCID})
}

}
