% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLSCV.R
\name{ConsensusOPLSCV}
\alias{ConsensusOPLSCV}
\title{ConsensusOPLSCV}
\usage{
ConsensusOPLSCV(
  K,
  Y,
  A,
  oax,
  nbrcv,
  cvType,
  preProcK = "no",
  preProcY = "no",
  cvFrac,
  modelType = "da",
  verbose = TRUE
)
}
\arguments{
\item{K}{matrix. The kernel matrix (un-centered); see \code{koplsKernel} for 
details.}

\item{Y}{matrix. The response matrix (un-centered/scaled). Could be binary 
(for discriminant analysis) or real-valued (for classical OPLS analysis).}

\item{A}{numeric. The number of Y-predictive components (integer).}

\item{oax}{numeric. The number of Y-orthogonal components (integer).}

\item{nbrcv}{numeric. Number of cross-validation rounds (integer).}

\item{cvType}{character. Type of cross-validation used. Either \code{nfold} 
for n-fold cross-validation, \code{mccv} for Monte Carlo cross-validation or 
\code{mccvb} for Monte Carlo class-balanced cross-validation. See also 
\code{koplsCrossValSet} for details.}

\item{preProcK}{character. Pre-processing settings for the kernel matrix. 
Either \code{mc} for mean-centering or \code{no} for no pre-processing.}

\item{preProcY}{character. Pre-processing parameter for Y. Either \code{mc} 
for mean-centering, \code{uv} for mean-centering and scaling to unit-variance, 
\code{pareto} for mean-centering and Pareto-scaling or \code{no} for no 
mean-centering and no scaling.}

\item{cvFrac}{numeric. Fraction of observations in the training set during 
cross-validation (integer).}

\item{verbose}{logical. Indicates whether the user wants to see the 
progress bar printed. If \code{FALSE}, no output will be printed. If 
\code{TRUE} (default) some output will be printed regarding the 
cross-validation progress.}

\item{modelType:}{character. Type of model used for the ConsensusOPLS method. 
It could be \code{da} for discriminant analysis or \code{reg} for regression. 
If \code{da} (default), sensitivity and specificity will be calculated.}
}
\value{

}
\description{
Function for performing Kernel-OPLS cross-validation for a set 
of Y-orthogonal components.
This function was first implemented in KOPLS1.1 MATLAB package, but was
modified in 2012 for the ConsensusOPLS method.
}
\examples{
#TO DO

}
\keyword{internal}
