% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLS-package.R
\docType{data}
\name{demo_3_Omics}
\alias{demo_3_Omics}
\title{Three-block omics data}
\description{
Three-block omics data
}
\references{
J. Boccard and D.N. Rutledge. A consensus OPLS-DA strategy for 
multiblock Omics data fusion. Analytica Chimica Acta, 769, 30-39, 2013.
}
\author{
Boccard & Rutledge
}
