% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsClassify.R
\name{koplsBasicClassify}
\alias{koplsBasicClassify}
\title{koplsBasicClassify (deprecated)}
\usage{
koplsBasicClassify(X, k)
}
\arguments{
\item{X}{matrix. It contains the predicted response matrix Y,
where columns denote classes and rows observations.}

\item{k}{numeric. Threshold value used to assign class categories.}
}
\value{
Predicted classification.
}
\description{
Classification function that assesses class belonging of a 
predicted response in \code{data} based on a fixed threshold \code{k}.
}
\keyword{internal}
