% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsCenterK.R
\name{koplsCenterKTeTe}
\alias{koplsCenterKTeTe}
\title{koplsCenterKTeTe}
\usage{
koplsCenterKTeTe(KteTe, KteTr, KtrTr)
}
\arguments{
\item{KteTe}{matrix. Contains the test kernel matrix, 
KteTe = <phi(Xte), phi(Xte)>.}

\item{KteTr}{matrix. Contains the hybrid test/training kernel matrix, 
KteTr = <phi(Xte), phi(Xtr)>.}

\item{KtrTr}{matrix. Contains the training kernel matrix, 
KtrTr = <phi(Xtr), phi(Xtr)>.}
}
\value{
\item{KteTe}{ matrix. The centered test kernel matrix.}
}
\description{
Centering function for the test kernel, which is constructed
from the test matrix Xte as KteTe = <phi(Xte), phi(Xte)>.
Requires additional (un-centered) kernels KteTr and KteTr to estimate mean 
values.

# ------------------------------------------------------------------------ #
This file is part of the K-OPLS package, developed by Max Bylesjo, 
University of Umea, Judy Fonville and Mattias Rantalainen, Imperial College.

Copyright (c) 2007-2010 Max Bylesjo, Judy Fonville and Mattias Rantalainen 

This code has been extended and adapted under the terms of the GNU General 
Public License version 2 as published by the Free Software Foundation.
# ------------------------------------------------------------------------ #
}
\examples{
KteTr <- matrix(1:25, nrow = 5, ncol = 5)
KteTe <- matrix(1:25, nrow = 5, ncol = 5)
KtrTr <- matrix(1:25, nrow = 5, ncol = 5)
test <- ConsensusOPLS:::koplsCenterKTeTe(KteTe = KteTe, KteTr = KteTr, KtrTr = KtrTr)
test

}
\keyword{internal}
