% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsCenterK.R
\name{koplsCenterKTeTr}
\alias{koplsCenterKTeTr}
\title{koplsCenterKTeTr}
\usage{
koplsCenterKTeTr(KteTr, KtrTr)
}
\arguments{
\item{KteTr}{matrix. Contains the hybrid test/training kernel matrix, 
KteTr = <phi(Xte), phi(Xtr)>.}

\item{KtrTr}{matrix. Contains the training kernel matrix; 
Ktrain = <phi(Xtr), phi(Xtr)>.}
}
\value{
The centered kernel matrix.
}
\description{
Centering function for the hybrid test/training kernel, which
is constructed from the test matrix Xte and the training matrix
Xtr as KteTr = <phi(Xte), phi(Xtr)>. Requires additional
(un-centered) training kernel to estimate mean values.
}
\examples{
Xte <- matrix(rnorm(20), ncol=5)
Xtr <- matrix(rnorm(25), ncol=5)
KteTr <- ConsensusOPLS:::koplsKernel(Xte, Xtr, Ktype='g', params=c(sigma=1.0))
KtrTr <- ConsensusOPLS:::koplsKernel(Xtr, Xtr, Ktype='g', params=c(sigma=1.0))
ConsensusOPLS:::koplsCenterKTeTr(KteTr = KteTr, KtrTr = KtrTr)

}
\keyword{internal}
