% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsCenterK.R
\name{koplsCenterKTrTr}
\alias{koplsCenterKTrTr}
\title{koplsCenterKTrTr}
\usage{
koplsCenterKTrTr(K)
}
\arguments{
\item{K}{matrix. Contains the training kernel matrix; K = <phi(Xtr), phi(Xtr)>.}
}
\value{
\item{K}{ matrix. The centered kernel matrix.}
}
\description{
Centering function for the training kernel, which is constructed
from the training matrix Xtr as K = <phi(Xtr), phi(Xtr)>.
}
\examples{
K <- matrix(1:25, nrow = 5, ncol = 5)
test <- ConsensusOPLS:::koplsCenterKTrTr(K = K)
test

}
\keyword{internal}
