% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsConfusionMatrix.R
\name{koplsConfusionMatrix}
\alias{koplsConfusionMatrix}
\title{koplsConfusionMatrix}
\usage{
koplsConfusionMatrix(trueClass, predClass)
}
\arguments{
\item{true_class}{vector. Indicates the true class belonging.}

\item{pred_class}{matrix. predicted class assignment.}
}
\value{
Confusion matrix.
}
\description{
Calculates a confusion matrix from classification results.
}
\examples{
trueClass <- c(1, 2, 1, 2, 3, 3)
predClass <- c(1, 2, 3, 2, 3, 3)
ConsensusOPLS:::koplsConfusionMatrix(trueClass = trueClass, 
                                     predClass = predClass)
}
\keyword{internal}
