% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsClassify.R
\name{koplsMaxClassify}
\alias{koplsMaxClassify}
\title{koplsMaxClassify}
\usage{
koplsMaxClassify(X)
}
\arguments{
\item{X}{matrix. It contains the predicted response matrix Y, where 
columns denote classes and rows observations.}
}
\value{
\item{predClass}{ matrix. It contains the predicted class(es) of \code{data}.}
}
\description{
Classification function that assesses class belonging of 'data' 
based on the maximum value.
}
\examples{
data <- as.matrix(data.frame(x1 = c(2, 1, 5, 1),
                             x2 = c(7, 1, 1, 5),
                             x3 = c(9, 5, 4, 9),
                             x4 = c(3, 4, 1, 2)))
test <- ConsensusOPLS:::koplsMaxClassify(X = data)
test

}
\keyword{internal}
