% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsModel.R
\name{koplsModel}
\alias{koplsModel}
\title{koplsModel}
\usage{
koplsModel(K, Y, A = 1, nox = 1, preProcK = "no", preProcY = "no")
}
\arguments{
\item{K}{matrix. Kernel matrix (un-centered); K = <phi(Xtr),phi(Xtr)>.}

\item{Y}{matrix. Response matrix (un-centered/scaled).}

\item{A}{numeric. Number of predictive components. Default is 1.}

\item{nox}{numeric. Number of Y-orthogonal components. Default is 1.}

\item{preProcK}{character. Pre-processing parameters for the 'K' matrix:
\code{mc} for mean-centering, \code{no} for no centering. Default is 'no'.}

\item{preProcY:}{character. Pre-processing parameters for the 'Y' matrix:
\code{mc} for mean-centering + no scaling, \code{uv} for mc + scaling to unit 
variance, \code{pa} for mc + scaling to Pareto, \code{no} for no centering + 
no scaling. Default is \code{no}.}
}
\value{
A list with the following entries:
\item{Cp}{ matrix. Y loading matrix.}
\item{Sp}{ matrix. Sigma matrix, containing singular values from Y'*K*Y used 
 for scaling.}
\item{Sps}{ matrix. Scaled Sigma matrix, containing scaled singular values.}
\item{Up}{ matrix. Y score matrix.}
\item{Tp}{ list. Predictive score matrix for all Y-orthogonal components.}
\item{T}{ matrix. Predictive score matrix for the final model.}
\item{co}{ list. Y-orthogonal loading vectors.}
\item{so}{ list. Eigenvalues from estimation of Y-orthogonal loading vectors.}
\item{to}{ list. Weight vector for the i-th latent component of the KOPLS 
model.}
\item{To}{ matrix. Y-orthogonal score matrix.}
\item{toNorm}{ list. Norm of the Y-orthogonal score matrix prior to scaling.}
\item{Bt}{ list. T-U regression coefficients for predictions.}
\item{A}{ numeric. Number of predictive components.}
\item{nox}{ numeric. Number of Y-orthogonal components.}
\item{K}{ matrix. The kernel matrix.}
\item{EEprime}{ matrix. The deflated kernel matrix for residual statistics.}
\item{sstot_K}{ numeric. Total sums of squares in 'K'.}
\item{R2X}{ numeric. Cumulative explained variation for all model components.}
\item{R2XO}{ numeric. Cumulative explained variation for Y-orthogonal model 
components.}
\item{R2XC}{ numeric. Explained variation for predictive model components 
after addition of Y-orthogonal model components. Ignored in this version.}
\item{sstot_Y}{ numeric. Total sums of squares in Y.}
\item{R2Y}{ numeric. Explained variation of Y.}
\item{R2Yhat}{ numeric. Variance explained by the i-th latent component of 
the model.}
\item{preProc$K}{ character. Pre-processing setting for K.}
\item{preProc$Y}{ character. Pre-processing setting for Y.}
\item{preProc$paramsY}{ character. Pre-processing scaling parameters for Y.}
}
\description{
Function for training a K-OPLS model. The function constructs a
predictive regression model for predicting the values of 'Y' by
using the information in 'K'. The explained variation is separated
into predictive components (dimensionality determined by the
parameter 'A') and 'Y'-orthogonal components (dimensionality determined 
by the parameter 'nox').
}
\examples{
K <- ConsensusOPLS:::koplsKernel(X1 = demo_3_Omics[["MetaboData"]], 
                                 X2 = NULL, type='p', params=c(order=1.0))
Y <- demo_3_Omics$Y
A <- 2
nox <- 4
preProcK <- "mc"
preProcY <- "mc"
model <- ConsensusOPLS:::koplsModel(K = K, Y = Y, A = A, nox = nox, 
                                   preProcK = preProcK, preProcY = preProcY)
ls(model)

}
\keyword{internal}
