% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsDummy.R
\name{koplsReDummy}
\alias{koplsReDummy}
\title{koplsReDummy}
\usage{
koplsReDummy(Y)
}
\arguments{
\item{Y}{matrix. A dummy matrix to be transformed into a numeric vector.}
}
\value{
The reconstructed integer class vector.
}
\description{
Reconstructs a (integer) class vector from a binary (dummy) 
matrix. This function is the inverse of \code{koplsDummy} function.
}
\examples{
class <- matrix(c(5, 1, 2, 3, 4, 3, 2, 4, 3, 1, 3), ncol = 1)
Y <- ConsensusOPLS:::koplsDummy(X = class)
X <- ConsensusOPLS:::koplsReDummy(Y)
X

}
\keyword{internal}
