% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsScale.R
\name{koplsScaleApply}
\alias{koplsScaleApply}
\title{koplsScaleApply}
\usage{
koplsScaleApply(model, X)
}
\arguments{
\item{model}{list. An object containing scaling parameters.}

\item{X}{matrix. The matrix to be scaled according to model parameters.}
}
\value{
A list containing the following entries:
\item{centerType}{ character. Indicates the centering method of the X matrix:
\code{mc} (mean-centering) or \code{no} (no centering).}
\item{scaleType}{ character. Indicates the scaling method of the X matrix: 
\code{uv} (unit variance), \code{pa} (pareto) or \code{no} (no scaling).}
\item{meanV}{ vector. Contains the mean values for all columns in X.}
\item{stdV}{ vector. Contains the standard deviations for all columns in X.}
\item{X}{ matrix. Scaled version of \code{varargin}, if defined. 
Otherwise, scaled version of \code{scaleS$X} from input. Scaling is done 
according to \code{centerType} and \code{scaleType}.
}
}
\description{
Applies scaling from external scaling objects on a matrix X.
}
\examples{
X <- matrix(data = c(1,4,7, 8,4,0, 3,6,9), nrow = 3)
Y <- ConsensusOPLS:::koplsScale(X = X, centerType = "mc", scaleType = "pa")
Z <- ConsensusOPLS:::koplsScaleApply(model = Y, X = Y$X)
Z$X

}
\keyword{internal}
