% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableImportanceProjection.R
\name{MBVIP}
\alias{MBVIP}
\title{MBVIP}
\usage{
MBVIP(data, Y, model = NULL, mc.cores = 1)
}
\arguments{
\item{data}{The collection list containing each block of data.}

\item{Y}{matrix. The response matrix to predict.}

\item{model}{A consensusOPLS model. Default, NULL, the model will be built.}

\item{mc.cores}{Number of cores for parallel computing. Default: 1.}
}
\value{
A table with the results:
\code{VIP} = sqrt(p*q/s), with:
\code{p} is the number of variables in each block
\code{q} is the explained variance of Y associated to each variable
\code{s} is the total Y variance explained by the model
}
\description{
Calculate the VIP (Variable Importance in Projection) for each variable of a 
Consensus-OPLS model.
}
\examples{
MBVIP(data=demo_3_Omics[c("MetaboData", "MicroData", "ProteoData")], 
      Y=demo_3_Omics$Y)
}
