% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLS.R
\docType{methods}
\name{plotContribution}
\alias{plotContribution}
\alias{plotContribution,ConsensusOPLS-method}
\title{Block contribution plot}
\usage{
plotContribution(object, col = NULL, ...)

\S4method{plotContribution}{ConsensusOPLS}(object, col = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ConsensusOPLS}.}

\item{col}{A vector of color codes or names, one for each block. Default,
NULL, 2 to number of blocks + 1.}

\item{...}{\code{barplot} arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Plot of relative contribution of each data block in the optimal model.
}
