% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLS.R
\name{plotDQ2}
\alias{plotDQ2}
\alias{plotDQ2,ConsensusOPLS-method}
\title{DQ2 plot}
\usage{
plotDQ2(
  object,
  breaks = 10,
  xlab = "DQ2",
  main = "DQ2 in models with permuted response",
  col = "blue",
  lty = "dashed",
  ...
)

\S4method{plotDQ2}{ConsensusOPLS}(
  object,
  breaks = 10,
  xlab = "DQ2",
  main = "DQ2 in models with permuted response",
  col = "blue",
  lty = "dashed",
  ...
)
}
\arguments{
\item{object}{An object of class \code{ConsensusOPLS}.}

\item{breaks}{See \code{hist}.}

\item{xlab}{See \code{hist}.}

\item{main}{See \code{hist}.}

\item{col}{A color code or name for DQ2 in the optimal model. Default, 2.
See \code{abline}.}

\item{lty}{A line type code or name for DQ2 in the optimal model. Default, 2.
See \code{abline}.}

\item{...}{\code{hist} arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Plot of DQ2 of models with permuted response.
}
