% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLS.R
\name{plotLoadings}
\alias{plotLoadings}
\alias{plotLoadings,ConsensusOPLS-method}
\title{Loading plot}
\usage{
plotLoadings(
  object,
  comp1 = "p_1",
  comp2 = "o_1",
  blockId = NULL,
  col = NULL,
  pch = NULL,
  ...
)

\S4method{plotLoadings}{ConsensusOPLS}(
  object,
  comp1 = "p_1",
  comp2 = "o_1",
  blockId = NULL,
  col = NULL,
  pch = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{ConsensusOPLS}.}

\item{comp1}{Latent variable for X-axis. Default, the first predictive
component, \code{p_1}.}

\item{comp2}{Latent variable for Y-axis. Default, the first orthogonal
component, \code{o_1}.}

\item{blockId}{The positions or names of the blocks for the plot.
Default, NULL, all.}

\item{col}{A vector of color codes or names, one for each block. Default,
NULL, 2 to \code{length(blockId)+1}.}

\item{pch}{A vector of graphic symbols, one for each block. Default, NULL,
1 to \code{length(blockId)}.}

\item{...}{\code{plot} arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Plot of variable loadings in the optimal model.
}
