% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLS.R
\name{plotQ2}
\alias{plotQ2}
\alias{plotQ2,ConsensusOPLS-method}
\title{Q2 plot}
\usage{
plotQ2(
  object,
  breaks = 10,
  xlab = "Q2",
  main = "Q2 in models with permuted response",
  col = "blue",
  lty = "dashed",
  ...
)

\S4method{plotQ2}{ConsensusOPLS}(
  object,
  breaks = 10,
  xlab = "Q2",
  main = "Q2 in models with permuted response",
  col = "blue",
  lty = "dashed",
  ...
)
}
\arguments{
\item{object}{An object of class \code{ConsensusOPLS}.}

\item{breaks}{See \code{\link{hist}}.}

\item{xlab}{See \code{\link{hist}}.}

\item{main}{See \code{\link{hist}}.}

\item{col}{A color code or name for Q2 in the optimal model. Default, 2.
See \code{\link{abline}}.}

\item{lty}{A line type code or name for Q2 in the optimal model. Default, 2.
See \code{\link{abline}}.}

\item{...}{\code{\link{hist}} arguments.}
}
\description{
Plot of Q2 of models with permuted response.

Plot of Q2 of models with permuted response.
}
