% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLS.R
\name{plotScores}
\alias{plotScores}
\alias{plotScores,ConsensusOPLS-method}
\title{Score plot}
\usage{
plotScores(object, comp1 = "p_1", comp2 = "o_1", col = NULL, pch = 19, ...)

\S4method{plotScores}{ConsensusOPLS}(object, comp1 = "p_1", comp2 = "o_1", col = NULL, pch = 19, ...)
}
\arguments{
\item{object}{An object of class \code{ConsensusOPLS}.}

\item{comp1}{Latent variable for abscissa. Default, the first predictive
component, \code{p_1}.}

\item{comp2}{Latent variable for ordinate. Default, the first orthogonal
component, \code{o_1}.}

\item{col}{A vector of color codes or names. Default, NULL, generated
following the \code{response}.}

\item{pch}{Graphic symbol. Default, 19.}

\item{...}{\code{plot} arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Plot of samples in the space of latent variables of the optimal model.
}
