% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLS.R
\name{plotVIP}
\alias{plotVIP}
\alias{plotVIP,ConsensusOPLS-method}
\title{VIP plot}
\usage{
plotVIP(
  object,
  comp1 = "p_1",
  comp2 = "p",
  blockId = NULL,
  col = NULL,
  pch = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)

\S4method{plotVIP}{ConsensusOPLS}(
  object,
  comp1 = "p_1",
  comp2 = "p",
  blockId = NULL,
  col = NULL,
  pch = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{ConsensusOPLS}.}

\item{comp1}{Latent variable for loadings on Y-axis. Default, the first
predictive component, \code{p_1}.}

\item{comp2}{Latent variable for VIPs on X-axis. Default, the predictive
component, \code{p}.}

\item{blockId}{The positions or names of the blocks for the plot.
Default, NULL, all.}

\item{col}{A vector of color codes or names, one for each block. Default,
NULL, 2 to \code{length(blockId)+1}.}

\item{pch}{A vector of graphic symbols, one for each block. Default, NULL,
1 to \code{length(blockId)}.}

\item{xlab}{X-axis label. Default, NULL, Loading on \code{comp}.}

\item{ylab}{Y-axis label. Default, NULL, VIP on \code{comp}.}

\item{...}{\code{plot} arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Plot of VIP versus variable loadings in the optimal model.
}
