% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLS.R
\name{predict}
\alias{predict}
\alias{predict,ConsensusOPLS-method}
\title{Model prediction}
\usage{
predict(object, newdata = NULL, nOcomp = NULL)

\S4method{predict}{ConsensusOPLS}(object, newdata = NULL, nOcomp = NULL)
}
\arguments{
\item{object}{An object of class \code{ConsensusOPLS}.}

\item{newdata}{A list of data frames of new data to predict. If omitted, the
fitted values in \code{object} are returned.}

\item{nOcomp}{Number of Y-orthogonal components to consider. Default, NULL,
the number of Y-orthogonal components in the optimal model.}
}
\description{
Predicts the response on new data with a fitted model.
}
