% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVConsensusOPLS.R
\name{RVConsensusOPLS}
\alias{RVConsensusOPLS}
\title{RVConsensusOPLS}
\usage{
RVConsensusOPLS(
  data,
  Y,
  A = 1,
  maxOrtholvs = 10,
  nrcv = 100,
  cvType = "nfold",
  modelType = "da",
  verbose = FALSE
)
}
\arguments{
\item{data}{The collection list containing each block of data.}

\item{Y}{The response matrix (un-centered/scaled).}

\item{A}{The number of Y-predictive components (integer).}

\item{maxOrtholvs}{The maximal number of Y-orthogonal components (integer).}

\item{nrcv}{Number of cross-validation rounds (integer).}

\item{cvType}{Type of cross-validation used. Either `nfold` for n-fold
cross-validation, `mccv` for Monte Carlo CV or `mccvb` for Monte Carlo 
class-balanced CV.}

\item{modelType}{type of OPLS regression model. Can be defined as "reg" for 
regression or "da" for discriminant analysis. Default value "da".}

\item{verbose}{logical which indicates whether the user wants to see the 
progress bar printlayed in the ConsensusOLPSCV function.}
}
\value{
`model`: a list with all model parameters.
}
\description{
Consensus OPLS-DA with RV coefficients weighting and 
DQ2 computation for discriminant analysis
}
\examples{
TO DO
}
