% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVConsensusOPLSPerm.R
\name{RVConsensusOPLSPerm}
\alias{RVConsensusOPLSPerm}
\title{RVConsensusOPLSPerm}
\usage{
RVConsensusOPLSPerm(data, Y, nbruns, PredLVs, maxOrtholvs)
}
\arguments{
\item{data}{list. The collection list containing each block of data.}

\item{Y}{matrix. The response matrix.}

\item{nbruns}{numeric. Number of random permutations.}

\item{PredLVs}{numeric. Number of predictive components.}

\item{maxOrtholvs}{numeric. Maximum number of orthogonal LVs to compute.}
}
\value{
A list with the two results plots: \code{plot_R2val}: Plot RV vs R2val.\code{plot_Q2val}: Plot RV vs Q2val.
}
\description{
Y random permutations for model validity estimation, with leave-on-out CV.
}
\examples{
data(demo_3_Omics)
RVConsensusOPLSPerm(data=demo_3_Omics[c("MetaboData", "MicroData", "ProteoData")], 
                    Y=demo_3_Omics$Y, nbruns=100, PredLVs=2, maxOrtholvs=2)
}
