% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVmodified.R
\name{RVmodified}
\alias{RVmodified}
\title{RVmodified}
\usage{
RVmodified(X, Y)
}
\arguments{
\item{X}{matrix. The normalized meta-kernel of all data block.}

\item{Y}{matrix. The centered and/or scaled kernel of each data block.}
}
\value{
The modified R-square value.
}
\description{
Calculate the coefficient of determination to assess how well the OPLS 
regression model fits the data. It indicates how close the observed values 
(real data) are to the values predicted by the regression model. Its 
calculation is modified for the present method, as it is adjusted over the 
range [0, 1].
}
\examples{
X <- matrix(stats::runif(n = 36), nrow=2)
Y <- matrix(stats::runif(n = 54), nrow=2)
result <- ConsensusOPLS:::RVmodified(X = X, Y = Y)
result

}
\keyword{internal}
