% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLS-package.R, R/ConsensusOPLS.R
\docType{package}
\name{ConsensusOPLS}
\alias{ConsensusOPLS}
\alias{_PACKAGE}
\alias{ConsensusOPLS-package}
\title{ConsensusOPLS}
\usage{
ConsensusOPLS(
  data,
  Y,
  maxPcomp = 1,
  maxOcomp = 5,
  modelType = "da",
  nperm = 100,
  cvType = "nfold",
  nfold = 5,
  nMC = 100,
  cvFrac = 4/5,
  kernelParams = list(type = "p", params = c(order = 1)),
  mc.cores = 1,
  plots = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{A list of data blocks. Each element of the list must be of matrix 
type. Rows and columns can have a name, in which case it will be retained 
during analysis. Any pre-processing of the data (e.g. scaling) must be 
carried out before building the list.}

\item{Y}{A vector, factor, dummy matrix or numerical matrix for the response. 
The type of answer given will condition the model to be used: a numerical 
vector for linear regression, a factor or dummy matrix for logistic 
regression or a discriminant model.}

\item{maxPcomp}{Maximum number of Y-predictive components. Default, 1.}

\item{maxOcomp}{Maximum number of Y-orthogonal components. Default, 5.}

\item{modelType}{String for type of OPLS regression model, either \code{reg} 
for regression or \code{da} for discriminant analysis. Default, \code{da}.}

\item{nperm}{Number of random permutations desired in response Y. Default, 100.}

\item{cvType}{String for type of cross-validation used. Either \code{nfold} 
for n-fold cross-validation, where \code{nfold} is look up, or \code{mccv} 
for Monte Carlo cross-validation, or \code{mccvb} for Monte Carlo 
class-balanced cross-validation, where \code{nMC} and \code{cvFrac} are used.
Default, \code{nfold}.}

\item{nfold}{Number of folds performed in n-fold cross-validation. This can 
be set to the number of samples to perform Leave-One-Out cross validation. 
Default, 5.}

\item{nMC}{An integer indicating the number of rounds performed when 
\code{cvType} is \code{mccv} or \code{mccvb}. Default, 100.}

\item{cvFrac}{A numeric value indicating the fraction of observations from 
\code{data} used in the training set for \code{mccv} or \code{mccvb} 
cross-validation. Default, 4/5 = 0.8.}

\item{kernelParams}{List of parameters for the kernel. Default, 
list(type='p', params = c(order=1.0)).}

\item{mc.cores}{Number of cores for parallel computing. Default, 1.}

\item{plots}{A logical indicating if plots are generated. For more aesthetic 
graphics, please refer to the package thumbnail. Interpretation help is also 
provided. Default, FALSE.}

\item{verbose}{A logical indicating if the computation progress will be shown. 
Default, FALSE.}
}
\value{
\code{ConsensusOPLS} returns a list of 
\item{\code{optimal}}{ results for optimal consensus OPLS model:}
\itemize{
     \item{Ys.}{ Response variable converted to dummy format.}
     \item{modelCV.}{ x.}
     \itemize{
         \item{Model.}{ x.}
         \itemize{
             \item{params.}{ Contains all model parameters such as number of 
             predictive components, number of orthogonal components, 
             sstot_K (numeric. Total sums of squares in \code{K}.), 
             sstot_Y ( numeric. Total sums of squares in Y.), 
             preProcK (character. Pre-processing setting for K.), 
             preProcY (character. Pre-processing setting for Y.), class.}
             \item{scoresP}{ Representation of the samples into the new 
             predictive components calculated by the optimal model.}
             \item{scoresO}{ Representation of the samples into the new 
             orthoginal components calculated by the optimal model.}
             \item{Cp}{ matrix. Y loading matrix.}
             \item{Sp}{ matrix. Sigma matrix, containing singular values from 
             \code{t(Y)* K *Y} used for scaling.}
             \item{Sps}{ matrix. Scaled Sigma matrix, containing scaled 
             singular values.}
             \item{Up}{ matrix. Y score matrix.}
             \item{Tp}{ list. Predictive score matrix for all Y-orthogonal 
             components.}
             \item{co}{ list. Y-orthogonal loading vectors.}
             \item{so}{ list. Eigenvalues from estimation of Y-orthogonal 
             loading vectors.}
             \item{to}{ list. Weight vector for the i-th latent component of 
             the KOPLS model.}
             \item{toNorm}{ list. Norm of the Y-orthogonal score matrix prior 
             to scaling.}
             \item{Bt}{ list. T-U regression coefficients for predictions.}
             \item{K}{ matrix. The kernel matrix.}
             \item{EEprime}{ matrix. The deflated kernel matrix for residual 
             statistics.}
             \item{R2X}{ numeric. Cumulative explained variation for all 
             model components.}
             \item{R2XO}{ numeric. Cumulative explained variation for 
             Y-orthogonal model components.}
             \item{R2Yhat}{ numeric. Variance explained by the i-th latent 
             component of the model.}
             \item{lambda}{ x.}
             \item{blockContribution}{ numeric. The specific importance of 
             each block to the latent predictor variable }
             \item{loadings}{ Individual loading of each block for the 
             predictive latent variable of the optimal model. They allow to 
             detect data level differences according to the Y response.}
             \item{scores}{ Representation of the samples into the new 
             components calculated by the optimal model.}
         }
         \item{cv}{ Cross-validation results.}
         \itemize{
             \item{AllYhat.}{ matrix. All predicted Y values as a 
             concatenated matrix for each block of data.}
             \item{Q2Yhat.}{ matrix. Total Q-square result for all 
             Y-orthogonal components.}
             \item{cvTestIndex.}{ matrix. Indices for the test set 
             observations during the cross-validation rounds.}
             \item{DQ2Yhat.}{ numeric. Discriminant Q2 value for all 
             Y-orthogonal components.}
             \item{nOcompOpt.}{ Number of orthogonal components used to build 
             the optimal model.}
         }
         \item{RV.}{ Value of the modified RV coefficient for each data block.}
         \item{normKernels.}{ Normalized kernel for each data block}
     }
     \item{VIP.}{ The variable Importance in projection (VIP) for each block of 
 data. Within each block, the relevance of the variables in explaining 
 variation in the Y response was assessed using the VIP parameter, which 
 reflects the importance of the variables in relation to both response and 
 projection quality.}
}
\item{\code{permuted}}{ models with permuted responses.}
\item{\code{permStats}}{ permutation statistics.}
\itemize{
     \item{lvnum}{ x.}
     \item{R2Yhat}{ numeric. Variance explained by the i-th latent component 
     of the model.}
     \item{DQ2Yhat}{ numeric. Discriminant Q2 value for all Y-orthogonal 
     components.}
     \item{Q2Yhat}{ matrix. Total Q-square result for all Y-orthogonal 
     components.}
     \item{Y}{ response variable in its dummy form.}
     \item{RV}{ Value of the modified RV coefficient for each data block.}
}
\item{\code{plots}}{ plots.}
}
\description{
A short description...

This package uses functions from the K-OPLS package, developed by Max Bylesjo, 
University of Umea, Judy Fonville and Mattias Rantalainen, Imperial College.

Copyright (c) 2007-2010 Max Bylesjo, Judy Fonville and Mattias Rantalainen 

This code has been extended and adapted under the terms of the GNU General 
Public License version 2 as published by the Free Software Foundation.

Constructs the consensus OPLS model with an optimal number of orthogonal 
components for given data blocks and Y response, and evaluate the model 
quality w.r.t other models built with randomly permuted responses.
}
\examples{
data(demo_3_Omics)
res <- ConsensusOPLS(data=demo_3_Omics[c("MetaboData", "MicroData", "ProteoData")], 
                     Y=demo_3_Omics$Y,
                     maxPcomp=1, maxOcomp=2, 
                     modelType='da',
                     nperm=5)
str(res)
}
\author{
\strong{Maintainer}: Van Du T. Tran \email{thuong.tran@sib.swiss} (\href{https://orcid.org/0000-0003-2074-5029}{ORCID})

Authors:
\itemize{
  \item Celine Bougel \email{celine.bougel@laposte.net} (\href{https://orcid.org/0000-0002-1598-7827}{ORCID})
  \item Julien Boccard \email{julien.boccard@unige.ch} (\href{https://orcid.org/0000-0001-5913-9566}{ORCID})
  \item Florence Mehl \email{florence.mehl@sib.swiss} (\href{https://orcid.org/0000-0002-9619-1707}{ORCID})
  \item Marie Tremblay-Franco \email{marie.tremblay-franco@inrae.fr} (\href{https://orcid.org/0000-0003-4678-4456}{ORCID})
}

}
