% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsClassify.R
\name{koplsBasicClassify}
\alias{koplsBasicClassify}
\title{koplsBasicClassify}
\usage{
koplsBasicClassify(X, k)
}
\arguments{
\item{X}{matrix. It contains the predicted response matrix Y,
where columns denote classes and rows observations.}

\item{k}{numeric. Threshold value used to assign class categories.}
}
\value{
\item{predClass}{ matrix. The predicted class(es) of \code{data} given 
\code{k}.}
}
\description{
Classification function that assesses class belonging of a 
predicted response in 'data' based on a fixed threshold 'k'.

# ------------------------------------------------------------------------ #
This file is part of the K-OPLS package, developed by Max Bylesjo, 
University of Umea, Judy Fonville and Mattias Rantalainen, Imperial College.

Copyright (c) 2007-2010 Max Bylesjo, Judy Fonville and Mattias Rantalainen 

This code has been extended and adapted under the terms of the GNU General 
Public License version 2 as published by the Free Software Foundation.
# ------------------------------------------------------------------------ #
}
\examples{
data <- as.matrix(data.frame(x1 = c(2, 1, 5, 1),
                             x2 = c(7, 1, 1, 5),
                             x3 = c(9, 5, 4, 9),
                             x4 = c(3, 4, 1, 2)))
k <- 4
test <- ConsensusOPLS:::koplsBasicClassify(X = data, k = k)
test

}
\keyword{internal}
