% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsClassify.R
\name{koplsBasicClassify}
\alias{koplsBasicClassify}
\title{koplsBasicClassify}
\usage{
koplsBasicClassify(X, k)
}
\arguments{
\item{X}{matrix. It contains the predicted response matrix Y,
where columns denote classes and rows observations.}

\item{k}{numeric. Threshold value used to assign class categories.}
}
\value{
Predicted classification.
}
\description{
Classification function that assesses class belonging of a 
predicted response in \code{data} based on a fixed threshold \code{k}.
}
\examples{
data <- as.matrix(data.frame(x1 = c(2, 1, 5, 1),
                             x2 = c(7, 1, 1, 5),
                             x3 = c(9, 5, 4, 9),
                             x4 = c(3, 4, 1, 2)))
k <- 4
test <- ConsensusOPLS:::koplsBasicClassify(X = data, k = k)
test

}
\keyword{internal}
