% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsCenterK.R
\name{koplsCenterKTeTe}
\alias{koplsCenterKTeTe}
\title{koplsCenterKTeTe}
\usage{
koplsCenterKTeTe(KteTe, KteTr, KtrTr)
}
\arguments{
\item{KteTe}{matrix. Contains the test kernel matrix, 
KteTe = <phi(Xte), phi(Xte)>.}

\item{KteTr}{matrix. Contains the hybrid test/training kernel matrix, 
KteTr = <phi(Xte), phi(Xtr)>.}

\item{KtrTr}{matrix. Contains the training kernel matrix, 
KtrTr = <phi(Xtr), phi(Xtr)>.}
}
\value{
The centered test kernel matrix.
}
\description{
Centering function for the test kernel, which is constructed
from the test matrix Xte as KteTe = <phi(Xte), phi(Xte)>.
Requires additional (un-centered) kernels KteTr and KteTr to estimate mean 
values.
}
\examples{
Xte <- matrix(data = stats::rnorm(n = 20), ncol=5)
Xtr <- matrix(data = stats::rnorm(n = 25), ncol=5)
KteTe <- ConsensusOPLS:::koplsKernel(X1 = Xte, X2 = Xte, 
                                     Ktype='p', params=c(order=1.0))
KteTr <- ConsensusOPLS:::koplsKernel(X1 = Xte, X2 = Xtr, 
                                     Ktype='p', params=c(order=1.0))
KtrTr <- ConsensusOPLS:::koplsKernel(X1 = Xtr, X2 = Xtr, 
                                     Ktype='p', params=c(order=1.0))
ConsensusOPLS:::koplsCenterKTeTe(KteTe = KteTe, 
                                 KteTr = KteTr, 
                                 KtrTr = KtrTr)
}
\keyword{internal}
