% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsCenterK.R
\name{koplsCenterKTeTr}
\alias{koplsCenterKTeTr}
\title{koplsCenterKTeTr}
\usage{
koplsCenterKTeTr(KteTr, KtrTr)
}
\arguments{
\item{KteTr}{matrix. Contains the hybrid test/training kernel matrix, 
KteTr = <phi(Xte), phi(Xtr)>.}

\item{KtrTr}{matrix. Contains the training kernel matrix; 
Ktrain = <phi(Xtr), phi(Xtr)>.}
}
\value{
\item{KteTr}{ matrix. The centered kernel matrix.}
}
\description{
Centering function for the hybrid test/training kernel, which
is constructed from the test matrix Xte and the training matrix
Xtr as KteTr = <phi(Xte), phi(Xtr)>. Requires additional
(un-centered) training kernel to estimate mean values.
}
\examples{
KteTr <- matrix(1:25, nrow = 5, ncol = 5)
KtrTr <- matrix(1:25, nrow = 5, ncol = 5)
test <- ConsensusOPLS:::koplsCenterKTeTr(KteTr = KteTr, KtrTr = KtrTr)
test

}
\keyword{internal}
