% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsCenterK.R
\name{koplsCenterKTrTr}
\alias{koplsCenterKTrTr}
\title{koplsCenterKTrTr}
\usage{
koplsCenterKTrTr(K)
}
\arguments{
\item{K}{matrix. Contains the training kernel matrix; K = <phi(Xtr), phi(Xtr)>.}
}
\value{
The centered kernel matrix.
}
\description{
Centering function for the training kernel, which is constructed
from the training matrix Xtr as K = <phi(Xtr), phi(Xtr)>.
}
\examples{
Xtr <- matrix(data = stats::rnorm(n = 25), ncol = 5)
K <- ConsensusOPLS:::koplsKernel(X1 = Xtr, X2 = Xtr, 
                                 Ktype='g', params=c(sigma=1.0))
ConsensusOPLS:::koplsCenterKTrTr(K = K)

}
\keyword{internal}
