% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsConfusionMatrix.R
\name{koplsConfusionMatrix}
\alias{koplsConfusionMatrix}
\title{koplsConfusionMatrix}
\usage{
koplsConfusionMatrix(true_class, pred)
}
\arguments{
\item{true_class}{vector. Indicates the true class belonging.}

\item{pred}{matrix. predicted class assignment.}
}
\value{
Confusion matrix.
}
\description{
Calculates a confusion matrix from classification results.
}
\examples{
true_class <- c(1, 2, 1, 2, 3, 3)
pred <- matrix(c(1, 2, 3, 2, 3, 3), nrow = length(true_class), ncol = 1)
test <- ConsensusOPLS:::koplsConfusionMatrix(true_class, pred)
test
}
