% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsCrossValSet.R
\name{koplsCrossValSet}
\alias{koplsCrossValSet}
\title{koplsCrossValSet}
\usage{
koplsCrossValSet(
  K,
  Y,
  modelFrac = 2/3,
  type = "nfold",
  nfold = NA,
  nfoldRound = NA
)
}
\arguments{
\item{K}{matrix. Kernel matrix.}

\item{Y}{matrix. Response matrix.}

\item{modelFrac}{numeric. Fraction (in percent) of observations used in the 
training set. By default is 2/3.}

\item{type}{character. Type of cross-validation wanted. It can be \code{nfold} 
for n-fold, \code{mccv} for Monte Carlo CV, \code{mccvb} for Monte Carlo 
class-balanced CV. Default is \code{nfold}.}

\item{nfold}{numeric. Number of total nfold rounds (if type = \code{nfold}).}

\item{nfoldRound}{numeric. Current nfold rounds (if type = \code{nfold}).}
}
\value{
A list with the following entries:
\item{CV_param}{ data frame. It contains \code{type} a character for the 
Cross-validation type, \code{nfold} the total number of nfold, 
\code{nfoldRound} the current nfold round and \code{class} the object class
wich is \code{crossValSet}.}
\item{KTrTr}{ matrix. Kernel training matrix; KTrTr = <phi(Xtr),phi(Xtr)>.}
\item{KTeTr}{ matrix. Kernel test/training matrix; KTeTr = <phi(Xte),phi(Xtr)>.}
\item{KTeTe}{ matrix. Kernel test matrix; KTeTe = <phi(Xte),phi(Xte)>.}
\item{yTraining}{ matrix. Y training set.}
\item{yTest}{ matrix. Y test set.}
\item{trainingIndex}{ integer. Indices of training set observations.}
\item{testIndex}{ numeric. Indices of test set observations.}
}
\description{
Generates set of training/test observations for the 
Cross-Validation.
}
\examples{
Y <- matrix(stats::rnorm(n = 100), nrow = 10)
K <- matrix(stats::rnorm(n = 100), nrow = 10)
type <- "nfold"  
modelFrac <- 0.7  
nfold <- 5  
nfoldRound <- 1 
test <- ConsensusOPLS:::koplsCrossValSet(K = K, Y = Y, 
                                         modelFrac = modelFrac,
                                         type = type, nfold = nfold, 
                                         nfoldRound = nfoldRound)
test 

}
\keyword{internal}
