% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsDummy.R
\name{koplsDummy}
\alias{koplsDummy}
\title{KoplsDummy}
\usage{
koplsDummy(X, numClasses = NA)
}
\arguments{
\item{X}{numeric vector with classes to define.}

\item{numClasses}{numeric. Pre-defined number of classes in the output.
By default, numClasses, number of unique values in \code{X}. Otherwise,
classes 1,...,numClasses are considered}
}
\value{
\item{X}{matrix. Results of the dummy matrix, with rows corresponding to
observations and columns to classes. Each element in matrix is either one
(observation belongs to class) or zero (observation does not belong to class).}
}
\description{
Transform a vector into a dummy (binary) matrix.
}
\examples{
class <- matrix(c(5, 1, 2, 3, 4, 3, 2, 4, 3, 1, 3), ncol = 1)
Y <- ConsensusOPLS:::koplsDummy(X = class, numClasses = NA)
Y

}
\keyword{internal}
