% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koplsScale.R
\name{koplsScale}
\alias{koplsScale}
\title{koplsScale}
\usage{
koplsScale(X, centerType = "no", scaleType = "no")
}
\arguments{
\item{X}{matrix. The matrix to be centered and scaled.}

\item{centerType}{character. Indicates the centering method of the \code{X} 
matrix: mean centering (\code{mc}) or no centering (\code{no}). Default is 
\code{no} centering.}

\item{scaleType}{character. Indicates the scaling method of the \code{X} 
matrix: \code{uv} for unit variance scaling (\code{pa}) for Pareto scaling or 
\code{no} for no scaling. Default is \code{no} scaling.}
}
\value{
A list containing the following entries:
\item{centerType}{ character. Indicates the centering method of the X matrix.}
\item{scaleType}{ character. Indicates the scaling method of the X matrix.}
\item{meanV}{ vector. Contains the mean values for all columns in X.}
\item{stdV}{ vector. Contains the standard deviations for all columns in X.}
\item{matrix}{ matrix. Original input matrix X, scaled according to 
\code{centerType} and \code{scaleType}.}
}
\description{
Function for mean-centering and scaling of a matrix.
}
\examples{
X <- matrix(c(1,4,7, 8,4,0, 3,6,9), nrow = 3)
Y <- ConsensusOPLS:::koplsScale(X, centerType = "mc", scaleType = "pa")
Y$matrix

}
\keyword{internal}
