% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusOPLS.R
\name{plotR2}
\alias{plotR2}
\alias{plotR2,ConsensusOPLS-method}
\title{R2 plot}
\usage{
plotR2(
  object,
  breaks = 10,
  xlab = "R2",
  main = "R2 in models with permuted response",
  col = "blue",
  lty = "dashed",
  ...
)

\S4method{plotR2}{ConsensusOPLS}(
  object,
  breaks = 10,
  xlab = "R2",
  main = "R2 in models with permuted response",
  col = "blue",
  lty = "dashed",
  ...
)
}
\arguments{
\item{object}{An object of class \code{ConsensusOPLS}.}

\item{breaks}{See \code{hist}.}

\item{xlab}{See \code{hist}.}

\item{main}{See \code{hist}.}

\item{col}{A color code or name for R2 in the optimal model. Default, 2.
See \code{abline}.}

\item{lty}{A line type code or name for R2 in the optimal model. Default, 2.
See \code{abline}.}

\item{...}{\code{hist} arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Plot of R2 of models with permuted response.
}
