% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVConsensusOPLS.R
\name{RVConsensusOPLS}
\alias{RVConsensusOPLS}
\title{RVConsensusOPLS}
\usage{
RVConsensusOPLS(
  data,
  Y,
  A = 1,
  maxOcomp = 5,
  nrcv = 3,
  cvType = "nfold",
  cvFrac = 2/3,
  modelType = "da",
  mc.cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{the collection list containing each block of data.}

\item{Y}{The response.}

\item{A}{The number of Y-predictive components (integer).}

\item{maxOcomp}{Maximum number of orthogonal components to compute.}

\item{nrcv}{Number of cross-validation rounds (integer).}

\item{cvType}{Type of cross-validation used. Either \code{nfold} for n-fold
cross-validation, \code{mccv} for Monte Carlo CV or \code{mccvb} for Monte 
Carlo class-balanced CV.}

\item{cvFrac}{numeric. Fraction of observations used in the training set. 
Default, 2/3. Fraction of data to be used for \code{mccv} and \code{mccvb}
cross-validation}

\item{modelType}{type of OPLS regression model. Can be defined as \code{reg} 
for regression or \code{da} for discriminant analysis. Default \code{da}.}

\item{mc.cores}{Number of cores for parallel computing. Default: 1.}

\item{verbose}{Logical which indicates whether the user wants to see the 
progress bar printed in the \code{ConsensusOLPSCV} function.}
}
\value{
A consensus OPLS model
}
\description{
Consensus OPLS-DA with RV coefficients weighting and 
DQ2 computation for discriminant analysis.
}
\examples{
data(demo_3_Omics)
ConsensusOPLS:::RVConsensusOPLS(data=demo_3_Omics[c("MetaboData", "MicroData", "ProteoData")], 
                                Y=demo_3_Omics$Y, modelType="da", A=1, mc.cores=1, nrcv=3)
}
\keyword{internal}
