% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVConsensusOPLS.R
\name{RVConsensusOPLS}
\alias{RVConsensusOPLS}
\title{RVConsensusOPLS}
\usage{
RVConsensusOPLS(
  data,
  Y,
  A = 1,
  maxOrtholvs = 10,
  nrcv = 10,
  cvType = "nfold",
  cvFrac = 2/3,
  modelType = "da",
  mc.cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{the collection list containing each block of data.}

\item{Y}{The response matrix (un-centered/scaled).}

\item{A}{The number of Y-predictive components (integer).}

\item{maxOrtholvs}{The maximal number of Y-orthogonal components (integer).}

\item{nrcv}{Number of cross-validation rounds (integer).}

\item{cvType}{Type of cross-validation used. Either \code{nfold} for n-fold
cross-validation, \code{mccv} for Monte Carlo CV or \code{mccvb} for Monte 
Carlo class-balanced CV.}

\item{cvFrac}{numeric. Fraction (in percent) of observations used in the 
training set. By default is 2/3. Fraction of data to be used for cross-validation}

\item{modelType}{type of OPLS regression model. Can be defined as \code{reg} 
for regression or \code{da} for discriminant analysis. Default value is
\code{da}.}

\item{mc.cores}{Number of cores for parallel computing. Default: 1.}

\item{verbose}{logical which indicates whether the user wants to see the 
progress bar printed in the \code{ConsensusOLPSCV} function.}
}
\value{
A consensus OPLS model
}
\description{
Consensus OPLS-DA with RV coefficients weighting and 
DQ2 computation for discriminant analysis.
}
\examples{
data(demo_3_Omics)
RVConsensusOPLS(data=demo_3_Omics[c("MetaboData", "MicroData", "ProteoData")], 
                Y=demo_3_Omics$Y, modelType="da", A=1, mc.cores=1, nrcv=1)
}
